html/user-activation/activation-trigger-keyup.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/testdriver.js"></script>
  <script src="/resources/testdriver-vendor.js"></script>
</head>
<body>
  <h1>Test for keyup activation trigger</h1>
  <p>Tests that a popup is allowed with user activation from a keyup event.</p>
  <input type="text" autofocus />
  <ol id="instructions">
    <li>Press ENTER key.
  </ol>
  <script>
    const ENTER_KEY = '\uE007';
    let test_w_enter = async_test("Popup with ENTER key");

    window.addEventListener("keyup", () => {
        test_w_enter.step(() => {
            let win = window.open();
            assert_true(!!win);
            win.close();
        });
        test_w_enter.done();
    });

    test_driver.send_keys(document.body, ENTER_KEY);
  </script>
</body>
</html>
